refactor: restructured project
